Fixed clock frequency calculation for STM32F4xx.
Changed the sequency of division and multiplication for the calculation of STM32_PLLVCO and STM32_PLLI2SVCO so that the result is integer also when not using the standard maximum settings.
Depending on clock configuration, STM32_PLLCLKIN can be inaccurate, if the result of STM32_HSECLK / STM32_PLLM_VALUE is not integer. However, the STM32_PLLCLKIN is kept in place so that the min/max checks continue to work. STM32_PLLCLKIN is not used for further calculation anymore though.